Move kubelet volume code to volumes.go
